reorg xmldsig1 tests into themed subtests
